Move index.ts outside of main, and rename main to app
