+++ client/src/index.scss
@@ -6,7 +6,7 @@ $BulmaLightWatermark: rgb(238, 237, 237);
 
 @import "globalStyles";
 
-@import "main/app";
+@import "app/app";
 @import "page/homepage";
 @import "page/search";
