+++ client/src/index.tsx
@@ -1,14 +1,14 @@
 import * as React from "react";
 import * as ReactDOM from "react-dom";
-import {Homepage} from "../page/homepage";
-import {App} from './app'
+import {Homepage} from "./page/homepage";
+import {App} from './app/app'
 // import Search from "./page.search";
-import {CreateProfile} from "../page/createProfile";
-import Login from "../page/login";
+import {CreateProfile} from "./page/createProfile";
+import Login from "./page/login";
 import {Provider, connect} from 'react-redux'
 import {Router, Route, browserHistory, IndexRoute} from 'react-router'
-import {store} from '../redux/store'
-import {requireUserLogin} from '../hoc/requireUserLogin'
+import {store} from './redux/store'
+import {requireUserLogin} from './hoc/requireUserLogin'
 import {syncHistoryWithStore, routerReducer} from 'react-router-redux'
 
 const history = syncHistoryWithStore(browserHistory, store)
