+++ client/src/webpackEntry.ts
@@ -1,3 +1,3 @@
 require('./index.scss')
 // require('material-design-lite/dist/material.min.js')
-import './main/index'
+import './index'
